Pagination with infinite scroll

+++ app/src/main/java/com/arctouch/codechallenge/api/TmdbApi.java
@@ -26,7 +26,7 @@ public interface TmdbApi {
     Observable<UpcomingMoviesResponse> upcomingMovies(
             @Query("api_key") String apiKey,
             @Query("language") String language,
-            @Query("page") Long page,
+            @Query("page") int page,
             @Query("region") String region
     );
